+++ src/Challenge/CodeChallengeResult.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ThreeBRS\EnterpriseSecurityBundle\Challenge;
+
+enum CodeChallengeResult
+{
+    case OK;
+    case EXPIRED;
+    case TOO_MANY_ATTEMPTS;
+    case INVALID;
+}